Split functions in other JS files
